move CommonChart to shared components
