+++ packages/contextual-compare/index.js
@@ -44,7 +44,6 @@ export default connect(
   mapDispatchToProps,
 )(ContextualCompare);
 
-export Chart from './components/Chart';
 export Title from './components/Title';
 // export reducer, actions and action types
 export reducer, { actions, actionTypes } from './reducer';
