+++ packages/shared-components/CommonChart/components/Footer/index.jsx
@@ -5,9 +5,7 @@ import {
   Text,
 } from '@bufferapp/components';
 
-import {
-  MetricIcon,
-} from '@bufferapp/analyze-shared-components';
+import MetricIcon from '../../../MetricIcon';
 
 const Metric = ({ metric }) => (
   <li
